leetcode

+++ 0150. Evaluate Reverse Polish Notation.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     int evalRPN(vector<string>& tokens) {
-        stack<int> st;
+        stack<long long> st;
         for(int i=0;i<tokens.size();i++){
             if(tokens[i]=="+" || tokens[i]=="-" || tokens[i]=="*" || tokens[i]=="/"){
-                int val1=st.top();st.pop();
-                int val2=st.top();st.pop();
+                long long val1=st.top();st.pop();
+                long long val2=st.top();st.pop();
                 switch(tokens[i][0]){
                     case '+': st.push(val2+val1);break;
                     case '-': st.push(val2-val1);break;
@@ -17,6 +17,6 @@ public:
                 st.push(stoi(tokens[i]));
             }
         }
-        return st.top();
+        return (int)st.top();
     }
 };
